+++ notebooks/utils.clj
@@ -66,7 +66,7 @@
 
 (defn sample-int
   [f ^long dx ^long dy]
-  (map (fn [v] [v (f v)]) (range dx (inc dy))))
+  (map (fn [v] [(+ v 0.5) (f v)]) (range dx (inc dy))))
 
 (defn bgraph-int
   ([f] (fgraph-int f nil))
@@ -75,7 +75,9 @@
   ([f domain rnge size]
    (let [[dx dy :as dom] (or domain [0 10])
          md (m/make-norm dx dy 5.0 (- size 5))
-         xsys (sample-int f dx dy)
+         xsys (if (map? f)
+                (seq f)
+                (sample-int f dx dy))
          [frx fry] (stats/extent (map second (filter (comp m/valid-double? second) xsys)))
          [rx ry] (case rnge
                    :domain dom
@@ -102,7 +104,7 @@
        (c2d/set-color c 0x303065)
        (doseq [[x y] xsys
                :when (m/valid-double? y)
-               :let [xx (md (+ x 0.5))]]
+               :let [xx (md x)]]
          (c2d/line c xx r0 xx (mr y)))
        (c2d/get-image c)))))
 
@@ -127,10 +129,28 @@
          n (m/make-norm v1 v2 0.0 1.0)]
      (c2d/with-canvas [c (c2d/canvas size size :high)]
        (c2d/set-background c 0xfafaff)
-       (c2d/set-color c 0x77303065)
        (doseq [[x y v] xy]
          (c2d/set-color c (grad (n v)))
          (c2d/crect c x y 1 1))
+       (c2d/get-image c)))))
+
+(defn scatter
+  ([xy] (scatter xy nil nil))
+  ([xy dx dy] (scatter xy dx dy size))
+  ([xy dx dy size]
+   (let [[dx1 dx2] (or dx [0.0 1.0])
+         [dy1 dy2] (or dy [0.0 1.0])
+         mx (m/make-norm dx1 dx2 5.0 (- size 5.0))
+         my (m/make-norm dy2 dy1 5.0 (- size 5.0))
+         d0 (mx 0.0)
+         r0 (my 0.0)]
+     (c2d/with-canvas [c (c2d/canvas size size :highest)]
+       (c2d/set-background c 0xfafaff)
+       (c2d/set-color c 0x77303065)
+       (c2d/line c 0 r0 size r0)
+       (c2d/line c d0 0 d0 size)
+       (doseq [[x y] xy]
+         (c2d/ellipse c (mx x) (my y) 3 3))
        (c2d/get-image c)))))
 
 (comment (graph2d r/noise))
@@ -143,9 +163,9 @@
 
 (defn dgraph-cont
   ([pdf-graph distr] (dgraph-cont pdf-graph distr nil))
-  ([pdf-graph distr {:keys [pdf icdf]
+  ([pdf-graph distr {:keys [pdf icdf data]
                      :or {pdf [0 1] icdf [0 0.999]}}]
-   (let [pdf-img (pdf-graph (partial r/pdf distr) pdf [0 nil] 100)
+   (let [pdf-img (pdf-graph (if data data (partial r/pdf distr)) pdf [0 nil] 100)
          cdf-img (fgraph-int (partial r/cdf distr) pdf [0 1] 100)
          icdf-img (fgraph-int (partial r/icdf distr) icdf :zero 100)]
      (c2d/with-canvas-> (c2d/canvas 310 120)
@@ -188,6 +208,7 @@
   ::clerk/viewer :hide-result}
 (comment
   (clerk/serve! {:browse? false :watch-paths ["notebooks"]})
+  (clerk/serve! {:browse? false})
   (clerk/show! 'nextjournal.clerk.tap)
   ;; (clerk/show! "notebooks/core.clj")
   (clerk/build! {:browse? false :paths ["notebooks/core.clj"